+++ components/OAuth.tsx
@@ -1,4 +1,4 @@
-import { Image, Text, View } from "react-native";
+/* import { Image, Text, View } from "react-native";
 import CustomButton from "./CustomButton";
 import { icons } from "@/constants";
 
@@ -31,3 +31,4 @@ const OAuth = () => {
 };
 
 export default OAuth;
+ */